Chatlog now officially connected with dbs

+++ src/main/java/nl/fontys/s3/domain/ChatDomains/UpdateChatStatusRequest.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @Builder
 public class UpdateChatStatusRequest {
-    int chatId;
+    long chatId;
     boolean status;
 }
